+++ mopidy/mpd/protocol/music_db.py
@@ -3,7 +3,6 @@ from __future__ import absolute_import, unicode_literals
 import functools
 import itertools
 
-from mopidy.internal import deprecation
 from mopidy.models import Track
 from mopidy.mpd import exceptions, protocol, translator
 
@@ -141,8 +140,7 @@ def find(context, *args):
     except ValueError:
         return
 
-    with deprecation.ignore('core.library.search:empty_query'):
-        results = context.core.library.search(query=query, exact=True).get()
+    results = context.core.library.search(query=query, exact=True).get()
     result_tracks = []
     if ('artist' not in query and
             'albumartist' not in query and
@@ -172,10 +170,9 @@ def findadd(context, *args):
 
     results = context.core.library.search(query=query, exact=True).get()
 
-    with deprecation.ignore('core.tracklist.add:tracks_arg'):
-        # TODO: for now just use tracks as other wise we have to lookup the
-        # tracks we just got from the search.
-        context.core.tracklist.add(tracks=_get_tracks(results)).get()
+    context.core.tracklist.add(
+        uris=[track.uri for track in _get_tracks(results)]
+    ).get()
 
 
 @protocol.commands.add('list')
@@ -439,8 +436,7 @@ def search(context, *args):
         query = _query_from_mpd_search_parameters(args, _SEARCH_MAPPING)
     except ValueError:
         return
-    with deprecation.ignore('core.library.search:empty_query'):
-        results = context.core.library.search(query).get()
+    results = context.core.library.search(query).get()
     artists = [_artist_as_track(a) for a in _get_artists(results)]
     albums = [_album_as_track(a) for a in _get_albums(results)]
     tracks = _get_tracks(results)
@@ -467,10 +463,9 @@ def searchadd(context, *args):
 
     results = context.core.library.search(query).get()
 
-    with deprecation.ignore('core.tracklist.add:tracks_arg'):
-        # TODO: for now just use tracks as other wise we have to lookup the
-        # tracks we just got from the search.
-        context.core.tracklist.add(_get_tracks(results)).get()
+    context.core.tracklist.add(
+        uris=[track.uri for track in _get_tracks(results)]
+    ).get()
 
 
 @protocol.commands.add('searchaddpl')